connected speech synthesis to whole codebase, still has to tweak some stuff

+++ src/tasks.py
@@ -48,15 +48,16 @@ def fetch_resource(url, path):
 
 
 @task
-def build(ctx, DTESTING="ON"):
+def build(ctx, DTESTING="ON", REMOVEBUILD="ON"):
     print("Building main project...")
 
     os.chdir(abs_path)
     print(f"Currently in {os.getcwd()} directory")
     print(DTESTING)
 
-    try: shutil.rmtree(join(abs_path, "build"))
-    except FileNotFoundError: pass
+    if REMOVEBUILD == "ON":
+        try: shutil.rmtree(join(abs_path, "build"))
+        except FileNotFoundError: pass
     ctx.run(f"cmake -B build -D TESTING={DTESTING}", pty=True)
     ctx.run("cmake --build build", pty=True)
 
@@ -65,7 +66,7 @@ def build(ctx, DTESTING="ON"):
 def run(ctx):
     print("Running main project...")
 
-    exec_directory = abs_path + "/build/main"
+    exec_directory = abs_path + "/build/test"
 
     ctx.run(exec_directory, pty=True)
